separate responder behaviour from sala agent
